feat(ui): add smooth fill easing to proximity-sidebar click
